refactor(subscription): remove unused components
refs #1061
